Create R Programming
When a US company wants to hire someone from outside of the United States for a technical
position, they have to file an application to the United States government to get a green card or visa
for the foreign applicant. These applications allow the US government to track who is entering and
leaving the country for work-related reasons, and ensure that immigrants are neither being taken
advantage of nor causing adverse effects for U.S. workers. To ensure equity for US and non-US
workers, companies have to state how much they are planning on paying the employee every time
they submit a visa or green card application. They also have to state the average amount an
employee with similar skills and background typically gets paid for the same position, a figure
called “the prevailing wage.” This publicly available data provides a unique view into what types
of salaries you might encounter for different data–related jobs in the US.